[TGBOT-389] Moving tests classes in correct directory and add doc for each class
