+++ rest-api-service/src/main/java/com/tutrit/storestapiservice/controller/CustomerController.java
@@ -1,6 +1,5 @@
 package com.tutrit.storestapiservice.controller;
 
-
 import com.tutrit.persistence.core.bean.Customer;
 import com.tutrit.storestapiservice.service.CustomerService;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,9 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestBody;
 
-
 @RestController
-public final class CustomerController {
+public class CustomerController {
 
     private final CustomerService customerService;
 
@@ -19,11 +17,23 @@ public final class CustomerController {
         this.customerService = customerService;
     }
 
+    /**
+     * Retrieves a customer by their ID.
+     *
+     * @param id The ID of the customer to retrieve.
+     * @return The customer with the specified ID.
+     */
     @GetMapping("/customers/{id}")
     public Customer getById(final @PathVariable String id) {
         return customerService.findById(id);
     }
 
+    /**
+     * Saves a new customer.
+     *
+     * @param customer The customer to be saved.
+     * @return The saved customer.
+     */
     @PostMapping("/customers")
     public Customer post(final @RequestBody Customer customer) {
         return customerService.save(customer);
