+++ rest-api-service/src/main/java/com/tutrit/storestapiservice/controller/OrderController.java
@@ -15,7 +15,7 @@ import javax.management.InstanceNotFoundException;
 import java.util.List;
 
 @RestController
-public final  class OrderController {
+public class OrderController {
 
     private final OrderService orderService;
 
@@ -23,34 +23,70 @@ public final  class OrderController {
         this.orderService = orderService;
     }
 
+    /**
+     * Retrieves all orders.
+     *
+     * @return A list of all orders.
+     */
     @GetMapping("/orders")
     public List<Order> findAll() {
         return orderService.findAll();
     }
 
+    /**
+     * Retrieves an order by its ID.
+     *
+     * @param id The ID of the order to retrieve.
+     * @return The order with the specified ID.
+     * @throws InstanceNotFoundException If the order
+     * with the specified ID is not found.
+     */
     @GetMapping("/orders/{id}")
-    public Order findAll(final @PathVariable String id)
-            throws InstanceNotFoundException {
+    public Order findAll(
+            final @PathVariable String id) throws InstanceNotFoundException {
         return orderService.findById(id);
     }
 
+    /**
+     * Saves a new order.
+     *
+     * @param order The order to be saved.
+     * @return The saved order.
+     */
     @PostMapping("/orders")
     public Order saveNew(final @RequestBody Order order) {
         return orderService.save(order);
     }
 
+    /**
+     * Updates an existing order.
+     *
+     * @param id    The ID of the order to update.
+     * @param order The updated order.
+     * @return The updated order.
+     * @throws AccessException If the provided
+     * order ID does not match the URI path.
+     */
     @PutMapping("/orders/{id}")
     public Order update(
             final @PathVariable String id,
-            final @RequestBody Order order) throws AccessException {
+            final @RequestBody Order order
+    ) throws AccessException {
         if (!id.equals(order.getOrderId())) {
-            throw new AccessException("Order uri path doesn't match order id");
-            // TODO : add custom exception handling
-
+            throw new AccessException(
+                    "Order URI path doesn't match order ID");
+            // TODO : Add custom exception handling
         }
         return orderService.save(order);
     }
 
+    /**
+     * Deletes an order by its ID.
+     *
+     * @param id The ID of the order to delete.
+     * @return true if the order was successfully
+     * deleted, false otherwise.
+     */
     @DeleteMapping("/orders/{id}")
     public boolean deleteById(final @PathVariable String id) {
         return orderService.delete(id);
