+++ rest-api-service/src/test/java/com/tutrit/storestapiservice/client/CarClientTest.java
@@ -2,6 +2,7 @@ package com.tutrit.storestapiservice.client;
 
 import com.tutrit.persistence.core.bean.Car;
 import com.tutrit.persistence.core.persistence.CarPersistence;
+import com.tutrit.storestapiservice.configurations.SpringContext;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,7 +11,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest
+@SpringBootTest(classes = SpringContext.SpringConfig.class)
 class CarClientTest {
     @Autowired
     private CarClient carClient;
