+++ rest-api-service/src/test/java/com/tutrit/storestapiservice/client/CustomerClientTest.java
@@ -1,7 +1,8 @@
-package com.tutrit.storestapiservice.controller;
+package com.tutrit.storestapiservice.client;
 
 import com.tutrit.persistence.core.bean.Customer;
-import com.tutrit.storestapiservice.service.CustomerService;
+import com.tutrit.persistence.core.persistence.CustomerPersistence;
+import com.tutrit.storestapiservice.configurations.SpringContext;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,28 +11,27 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest
-class CustomerControllerTest {
+@SpringBootTest(classes = SpringContext.SpringConfig.class)
+class CustomerClientTest {
     @Autowired
-    CustomerController customerController;
-
+    CustomerClient customerClient;
     @MockBean
-    CustomerService customerService;
-
+    CustomerPersistence customerPersistence;
 
     @Test
-    void getById() {
-        when(customerService.findById("1")).thenReturn(expectedCustomer());
-        Customer actualCustomer = customerController.getById("1");
+    void save() {
+        when(customerPersistence.save(changeCustomerBeforeSave())).thenReturn(expectedCustomer());
+        var actualCustomer = customerClient.save(createNewCustomer());
         assertEquals(expectedCustomer(), actualCustomer);
     }
 
-    @Test
-    void post() {
-        when(customerService.save(createNewCustomer())).thenReturn(expectedCustomer());
-        Customer actualCustomer = customerController.post(createNewCustomer());
-        assertEquals(expectedCustomer(), actualCustomer);
 
+    @Test
+    void findById() {
+
+        when(customerPersistence.findById("1")).thenReturn(expectedCustomer());
+        var actualCustomer = customerClient.findById("1");
+        assertEquals(expectedCustomer(), actualCustomer);
     }
 
     private Customer createNewCustomer() {
@@ -44,12 +44,17 @@ class CustomerControllerTest {
         return customer;
     }
 
+    private Customer changeCustomerBeforeSave() {
+        var customer = new Customer();
+        customer.setName("Yanki");
+        return customer;
+    }
 
     private Customer expectedCustomer() {
         var customer = new Customer();
-        customer.setName("Vasylyi");
+        customer.setName("Yanki");
         customer.setCustomerId("1");
-        customer.setCity("minsk");
+        customer.setCity("Brest");
         customer.setEmail("VasilBrest@mail.by");
         customer.setPhoneNumber("+375334573385");
         return customer;
