+++ rest-api-service/src/test/java/com/tutrit/storestapiservice/configurations/SpringContext.java
@@ -1,0 +1,16 @@
+package com.tutrit.storestapiservice.configurations;
+
+import org.springframework.boot.SpringBootConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+public class SpringContext {
+    @Configuration
+    @SpringBootConfiguration
+    @EnableAutoConfiguration
+    @ComponentScan(basePackages = "com.tutrit")
+    public static class SpringConfig {
+
+    }
+}
