+++ rest-api-service/src/test/java/com/tutrit/storestapiservice/service/CustomerServiceTest.java
@@ -1,7 +1,8 @@
-package com.tutrit.storestapiservice.client;
+package com.tutrit.storestapiservice.service;
 
 import com.tutrit.persistence.core.bean.Customer;
-import com.tutrit.persistence.core.persistence.CustomerPersistence;
+import com.tutrit.storestapiservice.client.CustomerClient;
+import com.tutrit.storestapiservice.configurations.SpringContext;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,26 +11,27 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest
-class CustomerClientTest {
+@SpringBootTest(classes = SpringContext.SpringConfig.class)
+class CustomerServiceTest {
+
     @Autowired
-    CustomerClient customerClient;
+    CustomerService customerService;
+
     @MockBean
-    CustomerPersistence customerPersistence;
+    CustomerClient customerClient;
 
     @Test
     void save() {
-        when(customerPersistence.save(changeCustomerBeforeSave())).thenReturn(expectedCustomer());
-        var actualCustomer = customerClient.save(createNewCustomer());
+        Customer customer = createNewCustomer();
+        when(customerClient.save(changeCustomerBeforeSave())).thenReturn(expectedCustomer());
+        Customer actualCustomer = customerService.save(customer);
         assertEquals(expectedCustomer(), actualCustomer);
     }
 
-
     @Test
     void findById() {
-
-        when(customerPersistence.findById("1")).thenReturn(expectedCustomer());
-        var actualCustomer = customerClient.findById("1");
+        when(customerClient.findById("1")).thenReturn(expectedCustomer());
+        var actualCustomer = customerService.findById("1");
         assertEquals(expectedCustomer(), actualCustomer);
     }
 
@@ -45,15 +47,15 @@ class CustomerClientTest {
 
     private Customer changeCustomerBeforeSave() {
         var customer = new Customer();
-        customer.setName("Yanki");
+        customer.setCity("minsk");
         return customer;
     }
 
     private Customer expectedCustomer() {
         var customer = new Customer();
-        customer.setName("Yanki");
+        customer.setName("Vasylyi");
         customer.setCustomerId("1");
-        customer.setCity("Brest");
+        customer.setCity("minsk");
         customer.setEmail("VasilBrest@mail.by");
         customer.setPhoneNumber("+375334573385");
         return customer;
